Improved vman() shell function (#13)
* Improved vman() shell function

This version of `vman()` checks to see if the requested man page exists before trying to launch superman.

This makes trying to look up a man page that doesn't exist faster and less visually jarring.

This PR assumes that `man` takes a superset of the arguments to `-c Superman`, which I'm not sure is the case. It appears to work with both 'vman 3 printf` and `vman printf`, although perhaps there are other forms which I don't know about.

PS Thanks for superman, it's rad!